refactor: 1시 이후 호출할 경우 baseDate문제 해결
